Impl Recipe & Fix Time to float

+++ src/de/fra/uas/digitales/kochbuch/backend/IRecipe.java
@@ -20,6 +20,6 @@ public interface IRecipe {
 
     int getRating();
 
-    int getTime();
+    float getTime();
 
 }
